Use vars to control time sync server options for ntp/chrony.
The settings for servers in the ntp.conf.j2 and chrony.conf.j2 templates
were previously hardcoded (to "iburst" and "minpoll 8", respectively),
but the user may wish to customise these.

This commit adds two new variables:
- rhel7cis_ntp_server_options
- rhel7cis_chrony_server_options

The default values are the same as the previously hardcoded values, so
there will be no change from the user side unless they override the
vars.